add patch route

+++ src/controllers/PostControllers.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 
 import { ERRORS } from "../consts";
 import { PostModel } from "../models";
-import { UserRequestWithId, PostCreate } from "../types";
+import { PostCreate, UserAuthorisedRequest, UserAuthorisedBodyRequest } from "../types";
 
 const getPosts = async (_req: Request, res: Response) => {
   try {
@@ -15,7 +15,7 @@ const getPosts = async (_req: Request, res: Response) => {
   }
 };
 
-const getPostById = async (req: Request, res: Response) => {
+const getPost = async (req: Request, res: Response) => {
   const _id = req.params?.id;
   try {
     const doc = await PostModel.findOneAndUpdate({ _id }, { $inc: { viewsCount: 1 } }, { returnDocument: "after" });
@@ -30,7 +30,7 @@ const getPostById = async (req: Request, res: Response) => {
   }
 };
 
-const createPost = async (req: UserRequestWithId<PostCreate>, res: Response) => {
+const createPost = async (req: UserAuthorisedBodyRequest<PostCreate>, res: Response) => {
   try {
     const post = await new PostModel({
       title: req.body.title,
@@ -47,10 +47,31 @@ const createPost = async (req: UserRequestWithId<PostCreate>, res: Response) => 
   }
 };
 
-const removePostById = async (req: Request, res: Response) => {
+const updatePost = async (req: UserAuthorisedBodyRequest<PostCreate>, res: Response) => {
   const _id = req.params?.id;
   try {
-    await PostModel.findOneAndDelete({ _id });
+    await PostModel.findOneAndUpdate(
+      { _id, user: { _id: req.userId } },
+      {
+        title: req.body.title,
+        text: req.body.text,
+        tags: req.body.tags,
+        imageUrl: req.body.imageUrl,
+        user: req.userId,
+      }
+    );
+
+    res.json({ success: true });
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({ message: ERRORS.SOMETHING_WENT_WRONG });
+  }
+};
+
+const removePost = async (req: UserAuthorisedRequest, res: Response) => {
+  const _id = req.params?.id;
+  try {
+    await PostModel.findOneAndDelete({ _id, user: { _id: req.userId } });
 
     res.json({ success: true });
   } catch (error) {
@@ -62,6 +83,7 @@ const removePostById = async (req: Request, res: Response) => {
 export const PostControllers = {
   createPost,
   getPosts,
-  getPostById,
-  removePostById,
+  getPost,
+  removePost,
+  updatePost,
 };
